Feature/selc 148

+++ core/src/main/java/it/pagopa/selfcare/product/core/ProductServiceImpl.java
@@ -1,12 +1,13 @@
 package it.pagopa.selfcare.product.core;
 
+import it.pagopa.selfcare.product.core.exception.ResourceNotFoundException;
 import it.pagopa.selfcare.product.dao.ProductRepository;
 import it.pagopa.selfcare.product.dao.model.Product;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.OffsetDateTime;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 class ProductServiceImpl implements ProductService {
@@ -25,44 +26,34 @@ class ProductServiceImpl implements ProductService {
 
     @Override
     public Product createProduct(Product product) {
-        return repository.save(new Product(product.getLogo(), product.getTitle(), product.getDescription(), product.getUrlPublic(), product.getUrlBO()));
-    }
-
-    @Override
-    public void deleteProducts() {
-        repository.deleteAll();
+        product.setActivationDateTime(OffsetDateTime.now());
+        return repository.save(product);
     }
 
     @Override
     public void deleteProduct(String id) {
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException();
+        }
     }
 
     @Override
     public Product getProduct(String id) {
-        Optional<Product> foundProduct = repository.findById(id);
-        Product result = null;
-        if (foundProduct.isPresent()) {
-            result = foundProduct.get();
-        }
-        return result;
+        return repository.findById(id).orElseThrow(ResourceNotFoundException::new);
     }
 
 
     @Override
     public Product updateProduct(String id, Product product) {
-        Optional<Product> foundProduct = repository.findById(id);
-        Product result = null;
-        if (foundProduct.isPresent()) {
-            Product p = foundProduct.get();
-            p.setLogo(product.getLogo());
-            p.setTitle(product.getTitle());
-            p.setDescription(product.getDescription());
-            p.setUrlPublic(product.getUrlPublic());
-            p.setUrlBO(product.getUrlBO());
-            result = repository.save(p);
-        }
-        return result;
+        Product foundProduct = repository.findById(id).orElseThrow(ResourceNotFoundException::new);
+        foundProduct.setLogo(product.getLogo());
+        foundProduct.setTitle(product.getTitle());
+        foundProduct.setDescription(product.getDescription());
+        foundProduct.setUrlPublic(product.getUrlPublic());
+        foundProduct.setUrlBO(product.getUrlBO());
+        return repository.save(foundProduct);
     }
 
 }
